refator: 코드리펙토링 및 분리
