+++ com.spring-cloud.eureka.client.auth/src/main/java/com/spring_cloud/eureka/client/auth/application/security/JwtUtil.java
@@ -13,6 +13,7 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 import org.springframework.util.StringUtils;
 
+import javax.crypto.SecretKey;
 import java.security.Key;
 import java.util.Base64;
 import java.util.Date;
